application/src/api/requirements.txt
# FastAPI Backend for EA Agentic Lab iOS App
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pyyaml>=6.0
httpx>=0.26.0
python-dotenv>=1.0.0
